@@ -5,11 +5,12 @@ ruby "2.6.10"
 gem 'rake'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '5.1'
+gem 'rails', '5.2'
 # Use SCSS for stylesheets
 gem 'sass-rails', '~> 5'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
+gem "bootsnap", require: false
 
 # Use jquery as the JavaScript library
 gem 'jquery-rails'
@@ -24,7 +25,7 @@ gem 'sdoc', '~> 0.4.0',          group: :doc
 # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
 gem 'spring',        group: :development
 
-gem 'pg', '~> 0.18'
+gem 'pg'
 
 gem 'bootstrap-sass', '~> 3.2.0.4'
 gem 'autoprefixer-rails'
